+++ .gitignore
@@ -1,0 +1,2 @@
+.jpg
+__pycache__
