+++ environment.yml
@@ -1,0 +1,61 @@
+name: xhs_publisher
+channels:
+  - defaults
+  - https://repo.anaconda.com/pkgs/main
+  - https://repo.anaconda.com/pkgs/r
+dependencies:
+  - bzip2=1.0.8=h6c40b1e_6
+  - ca-certificates=2025.2.25=hecd8cb5_0
+  - expat=2.6.4=h6d0c2b6_0
+  - libcxx=14.0.6=h9765a3e_0
+  - libffi=3.4.4=hecd8cb5_1
+  - ncurses=6.4=hcec6c5f_0
+  - openssl=3.0.16=h184c1cd_0
+  - pip=25.0=py312hecd8cb5_0
+  - python=3.12.8=hcd54a6c_0
+  - readline=8.2=hca72f7f_0
+  - setuptools=75.8.0=py312hecd8cb5_0
+  - sqlite=3.45.3=h6c40b1e_0
+  - tk=8.6.14=h4d00af3_0
+  - tzdata=2025a=h04d1e81_0
+  - wheel=0.45.1=py312hecd8cb5_0
+  - xz=5.6.4=h46256e1_1
+  - zlib=1.2.13=h4b97444_1
+  - pip:
+      - altgraph==0.17.4
+      - attrs==25.3.0
+      - certifi==2025.1.31
+      - charset-normalizer==3.4.1
+      - click==8.1.8
+      - greenlet==3.0.3
+      - h11==0.14.0
+      - idna==3.10
+      - macholib==1.16.3
+      - outcome==1.3.0.post0
+      - packaging==24.2
+      - pillow==11.1.0
+      - playwright==1.46.0
+      - pydantic==1.10.13
+      - pyee==11.1.0
+      - pyinstaller==6.10.0
+      - pyinstaller-hooks-contrib==2025.2
+      - pyqt6==6.8.1
+      - pyqt6-qt6==6.8.2
+      - pyqt6-sip==13.10.0
+      - pysocks==1.7.1
+      - python-dotenv==1.1.0
+      - requests==2.32.3
+      - selenium==4.27.1
+      - sniffio==1.3.1
+      - sortedcontainers==2.4.0
+      - trio==0.29.0
+      - trio-websocket==0.12.2
+      - typing-extensions==4.13.0
+      - undetected-chromedriver==3.5.5
+      - urllib3==2.3.0
+      - uvicorn==0.34.0
+      - webdriver-manager==4.0.1
+      - websocket-client==1.8.0
+      - websockets==15.0.1
+      - wsproto==1.2.0
+prefix: /Users/smallshaxiaoche/miniconda3/envs/xhs_publisher
