+++ src/core/xiaohongshu_cookies.json
@@ -1,0 +1,1 @@
+[{"name": "acw_tc", "value": "0a0d068317430465580874381e0b025328c45d6764a1f0a9c89da6ef1da138", "domain": "creator.xiaohongshu.com", "path": "/", "expires": 1743048357.845124, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "xsecappid", "value": "ugc", "domain": ".xiaohongshu.com", "path": "/", "expires": 1774582613, "httpOnly": false, "secure": false, "sameSite": "Lax"}, {"name": "a1", "value": "195d5abc1cbh3ue49h216kxkaute8hx4c8knnefd730000211548", "domain": ".xiaohongshu.com", "path": "/", "expires": 1774582558, "httpOnly": false, "secure": false, "sameSite": "Lax"}, {"name": "webId", "value": "21dbca514134b075d73adbab7f4987dc", "domain": ".xiaohongshu.com", "path": "/", "expires": 1774582558, "httpOnly": false, "secure": false, "sameSite": "Lax"}, {"name": "acw_tc", "value": "0a00d93117430465586058887e58157371f107e4c046d6ce9dbf3cc386a258", "domain": "www.xiaohongshu.com", "path": "/", "expires": 1743048358.355976, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "acw_tc", "value": "0a4a61c717430465586228080e60e6215e2326e89f9d2d2fcb7fefcd866fd6", "domain": "edith.xiaohongshu.com", "path": "/", "expires": 1743048358.365332, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "acw_tc", "value": "0a0d0ff917430465587374402e093d297cec1fa4a6e8cacb7f0c81a4643ccf", "domain": "customer.xiaohongshu.com", "path": "/", "expires": 1743048358.560265, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "gid", "value": "yj2f20DfqWSWyj2f20DSyYdiSDxq7d4jxJyKTfWTCT079Iq8dYxC4I888Jyy24Y8if2d2jfd", "domain": ".xiaohongshu.com", "path": "/", "expires": 1777606562.434897, "httpOnly": false, "secure": false, "sameSite": "Lax"}, {"name": "customer-sso-sid", "value": "68c517486328198656262085uhdnwax7xpkb9rkb", "domain": ".xiaohongshu.com", "path": "/", "expires": 1743651412.002869, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "x-user-id-creator.xiaohongshu.com", "value": "58629aa26a6a697e60d845e4", "domain": ".xiaohongshu.com", "path": "/", "expires": 1777606613.003763, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "customerClientId", "value": "467585921229809", "domain": ".xiaohongshu.com", "path": "/", "expires": 1777606613.003926, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "access-token-creator.xiaohongshu.com", "value": "customer.creator.AT-68c517486328198656262086jijrbz71zvxqzrv5", "domain": ".xiaohongshu.com", "path": "/", "expires": 1745638612.003996, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "galaxy_creator_session_id", "value": "CE9FSxchnEJzjyzSa9BiFgyWAWLBM8vlNz1n", "domain": ".xiaohongshu.com", "path": "/", "expires": 1745638613.004047, "httpOnly": true, "secure": false, "sameSite": "Lax"}, {"name": "galaxy.creator.beaker.session.id", "value": "1743046613179060847222", "domain": ".xiaohongshu.com", "path": "/", "expires": 1745638613.004092, "httpOnly": true, "secure": false, "sameSite": "Lax"}]
